Changed memory access format * 8 bits only now, records 255 when memory touched at current frame, reduced to 200 at end of frame, then subtract 1 at subsequent frames * breakpoint will highlight current access with very bright 255, otherwise 200 is max shown

+++ libatari800/atari800_bridge.h
@@ -39,7 +39,7 @@ typedef struct {
         uint8_t unused1;
         uint8_t unused2;
 
-        uint16_t memory_access[MAIN_MEMORY_SIZE];
+        uint8_t memory_access[MAIN_MEMORY_SIZE];
         uint8_t access_type[MAIN_MEMORY_SIZE];
 
         uint8_t PC[2];
